moved utility functions code from the folder responsible for routing to a different folder for code cleanliness
